+++ tests/testthat/test-trace_calls.R
@@ -6,15 +6,15 @@ test_that("one-line functions are traced correctly", {
 
   fun <- function(x) x + 1
 
-  expect_equal(as.character(body(trace_calls(fun))[[2]][[1]]),
+  expect_equal(as.character(body(trace_calls(fun))[[2]][[2]][[1]]),
       c(":::", "covr", "count"))
 
   fun <- function() 1
 
-  expect_equal(as.character(body(trace_calls(fun))[[2]][[1]]),
+  expect_equal(as.character(body(trace_calls(fun))[[2]][[2]][[1]]),
       c(":::", "covr", "count"))
 
-  expect_equal(body(trace_calls(fun))[[3]], body(fun))
+  expect_equal(body(trace_calls(fun))[[2]][[3]], body(fun))
 })
 test_that("one-line functions with no calls are traced correctly", {
   old <- getOption("keep.source")
@@ -23,10 +23,10 @@ test_that("one-line functions with no calls are traced correctly", {
 
   fun <- function(x) x
 
-  expect_equal(as.character(body(trace_calls(fun))[[2]][[1]]),
+  expect_equal(as.character(body(trace_calls(fun))[[2]][[2]][[1]]),
       c(":::", "covr", "count"))
 
-  expect_equal(body(trace_calls(fun))[[3]], body(fun))
+  expect_equal(body(trace_calls(fun))[[2]][[3]], body(fun))
 })
 test_that("one-line functions with braces are traced correctly", {
   old <- getOption("keep.source")
@@ -37,10 +37,10 @@ test_that("one-line functions with braces are traced correctly", {
     x + 1
   }
 
-  expect_equal(as.character(body(trace_calls(fun))[[2]][[2]][[1]]),
+  expect_equal(as.character(body(trace_calls(fun))[[2]][[2]][[2]][[1]]),
       c(":::", "covr", "count"))
 
-  expect_equal(body(trace_calls(fun))[[2]][[3]], body(fun)[[2]])
+  expect_equal(body(trace_calls(fun))[[2]][[2]][[3]], body(fun)[[2]])
 })
 
 test_that("one-line functions with no calls and braces are traced correctly", {
@@ -52,7 +52,7 @@ test_that("one-line functions with no calls and braces are traced correctly", {
     1
   }
 
-  e2 <- body(trace_calls(fun))[[2]]
+  e2 <- body(trace_calls(fun))[[2]][[2]]
   expect_true(length(e2) > 1 &&
               identical(as.character(e2[[2]][[1]]), c(":::", "covr", "count")))
 
@@ -60,7 +60,7 @@ test_that("one-line functions with no calls and braces are traced correctly", {
     x
   }
 
-  e2 <- body(trace_calls(fun))[[2]]
+  e2 <- body(trace_calls(fun))[[2]][[2]]
 
   # the second expr should be a block
   expect_true(length(e2) > 1 &&
@@ -85,7 +85,7 @@ test_that("last evaled expression is traced", {
   # last expression: the implicit return expression
   e3 <- body[[3]]
   expect_true(length(e3) > 1 &&
-    identical(as.character(e3[[2]][[1]]), c(":::", "covr", "count")))
+    identical(as.character(e3[[2]][[2]][[1]]), c(":::", "covr", "count")))
 
 })
 
@@ -96,4 +96,8 @@ test_that("functions with NULL bodies are traced correctly", {
   fun <- function() NULL
 
   expect_null(trace_calls(fun)())
+})
+
+test_that("functions with curly curly syntax are traced correctly", {
+  #TODO: good tests for this
 })